Overview Demo refactoring
