+++ src/Demos/OverviewDemo/OverviewDemo.tsx
@@ -3,104 +3,10 @@ import './OverviewDemo.scss';
 import React, { useState } from 'react';
 
 import { ITableOption, Table } from '../../lib';
-import {
-  ActionType, DataType, EditingMode, FilteringMode, SortDirection, SortingMode,
-} from '../../lib/enums';
-import { Cell } from '../../lib/models';
-import {
-  CellFuncPropsWithChildren, DataChangeFunc, FilterRowFuncPropsWithChildren, OptionChangeFunc,
-} from '../../lib/types';
-import { dateUtils } from '../../lib/utils';
+import { DataType, EditingMode, FilteringMode, SortDirection, SortingMode } from '../../lib/enums';
+import { DataChangeFunc, OptionChangeFunc } from '../../lib/types';
 import dataArray from './data';
-
-const CustomCell: React.FC<CellFuncPropsWithChildren> = ({
-  column: {
-    key,
-  },
-  dispatch,
-  rowData,
-  rowKeyField,
-  value,
-}) => {
-  return (
-    <div onClick={() => {
-      const cell: Cell = { columnKey: key, rowKey: rowData[rowKeyField] };
-      dispatch(ActionType.OpenEditor, { cell });
-    }}>
-      {value ? 'Passed' : 'Failed'}
-    </div>
-  );
-};
-
-const CustomImageCell: React.FC<CellFuncPropsWithChildren> = ({
-  value,
-}) => {
-  return (
-    <div>
-      <img className='custom-cell-image' src={value} alt=''/>
-    </div>
-  );
-};
-
-const FilterOperators: React.FC<FilterRowFuncPropsWithChildren> = ({
-  column, dispatch,
-}) => {
-  return (
-    <select
-      className='form-control'
-      defaultValue={column.filterRowOperator}
-      onChange={(event) => {
-        dispatch(ActionType.ChangeFilterRow, { column: {...column, filterRowOperator: event.currentTarget.value}});
-      }}>
-      <option value={'='}>=</option>
-      <option value={'<'}>{'<'}</option>
-      <option value={'>'}>{'>'}</option>
-      <option value={'<='}>{'<='}</option>
-      <option value={'>='}>{'>='}</option>
-    </select>
-  );
-};
-
-const CustomNumberFilterEditor: React.FC<FilterRowFuncPropsWithChildren> = ({
-  column, dispatch,
-}) => {
-  return (
-    <div>
-      <FilterOperators column={column} dispatch={dispatch}/>
-      <input
-        defaultValue={column.filterRowValue}
-        style={{width: 60}}
-        onChange={(event) => {
-          const filterRowValue = event.currentTarget.value !== '' ? Number(event.currentTarget.value) : null;
-          dispatch(ActionType.ChangeFilterRow, { column: {...column, filterRowValue}});
-        }}
-        type='number'
-      />
-    </div>
-  );
-};
-
-const CustomDateFilterEditor: React.FC<FilterRowFuncPropsWithChildren> = ({
-  column, dispatch,
-}) => {
-  const fieldValue = column.filterRowValue;
-  const value = fieldValue && dateUtils.getDateInputValue(fieldValue);
-  return (
-    <div>
-      <FilterOperators column={column} dispatch={dispatch}/>
-      <input
-        type='date'
-        value={value || ''}
-        onChange={(event) => {
-          const targetValue = event.currentTarget.value;
-          const filterRowValue = targetValue ? new Date(targetValue) : null;
-          const updatedColumn = {...column, filterRowValue};
-          dispatch(ActionType.ChangeFilterRow, {column: updatedColumn});
-        }}
-      />
-    </div>
-  );
-};
+import { CustomDateFilterEditor, CustomImageCell, CustomNumberFilterEditor } from './Editors';
 
 const tableOption: ITableOption = {
   columns: [
@@ -122,7 +28,6 @@ const tableOption: ITableOption = {
       title: 'Representative',
     },
     {
-      cell: CustomCell,
       dataType: DataType.Boolean,
       fieldParents: ['company'],
       format: (value) => `Loyal program: ${ value ? 'Yes' : 'No'}`,
@@ -181,17 +86,11 @@ const tableOption: ITableOption = {
       filterRowCell: CustomDateFilterEditor,
       filterRowOperator: '<',
       filterRowValue: new Date(2015, 1, 2),
-      format: (value: Date) => {
-        const ageDifMs = Date.now() - value.getTime();
-        const ageDate = new Date(ageDifMs);
-        const count = Math.abs(ageDate.getUTCFullYear() - 1970);
-        const dateString = value.toLocaleDateString('en', {
-          day: '2-digit',
-          month: '2-digit',
-          year: 'numeric',
-        });
-        return `${dateString} (${count} year${count === 1 ? '' : 's' } ago)`;
-      },
+      format: (value: Date) => value.toLocaleDateString('en', {
+        day: '2-digit',
+        month: '2-digit',
+        year: 'numeric',
+      }),
       key: 'firstDealDate1',
       style: {
         textAlign: 'center',
